refactor(tools): extract tool_type, tool_config, tool_product into dojo/<module>/ [Phase 1,3,5,6,8,9]
