+++ dojo/tool_config/__init__.py
@@ -1,0 +1,1 @@
+import dojo.tool_config.admin  # noqa: F401
