+++ dojo/tool_config/api/__init__.py
@@ -1,0 +1,1 @@
+path = "tool_configurations"  # noqa: RUF067
