+++ dojo/tool_config/api/urls.py
@@ -1,0 +1,6 @@
+from dojo.tool_config.api.views import ToolConfigurationsViewSet
+
+
+def add_tool_config_urls(router):
+    router.register(r"tool_configurations", ToolConfigurationsViewSet, basename="tool_configuration")
+    return router
