+++ dojo/tool_product/__init__.py
@@ -1,0 +1,1 @@
+import dojo.tool_product.admin  # noqa: F401
