+++ dojo/tool_product/api/__init__.py
@@ -1,0 +1,1 @@
+path = "tool_product_settings"  # noqa: RUF067
