+++ dojo/tool_product/signals.py
@@ -3,8 +3,8 @@ import logging
 from django.db.models.signals import pre_delete
 from django.dispatch import receiver
 
-from dojo.models import Tool_Product_Settings
 from dojo.notes.helper import delete_related_notes
+from dojo.tool_product.models import Tool_Product_Settings
 
 logger = logging.getLogger(__name__)
