+++ dojo/tool_type/__init__.py
@@ -1,0 +1,1 @@
+import dojo.tool_type.admin  # noqa: F401
