+++ dojo/tool_type/admin.py
@@ -1,0 +1,5 @@
+from django.contrib import admin
+
+from dojo.tool_type.models import Tool_Type
+
+admin.site.register(Tool_Type)
