+++ dojo/tool_type/api/__init__.py
@@ -1,0 +1,1 @@
+path = "tool_types"  # noqa: RUF067
